requirements.txt
flask>=3.0
python-dotenv>=1.0
itsdangerous>=2.0
gunicorn>=21.0
